--- src/components/CartIconWithCount/index.tsx
@@ -1,1 +1,0 @@
-export { default } from './CartIconWithCount';
