+++ src/components/CartStepperWithIcon/CartStepperWithIcon.style.ts
