+++ src/components/ExpectedPayment/index.ts
@@ -1,0 +1,1 @@
+export { default } from './ExpectedPayment';
