--- src/components/PayingContainer/index.ts
@@ -1,1 +1,0 @@
-export { default } from './PayingContainer';
