--- src/components/common/QuantityCounter/index.tsx
@@ -1,1 +1,0 @@
-export { default } from './QuantityCounter';
